+++ start/src/main/webapp/error403.html
@@ -1,0 +1,44 @@
+<!--
+  Copyright (c) 2018 IBM Corp.
+
+  Licensed under the Apache License, Version 2.0 (the "License");
+  you may not use this file except in compliance with the License.
+  You may obtain a copy of the License at
+
+     http://www.apache.org/licenses/LICENSE-2.0
+
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+-->
+<html>
+<head>
+    <title>Authorization Failed</title>
+    <style>
+      #logoutButton {
+          width: 100%;
+          border-bottom: 1px solid black;
+          margin-bottom: 1em;
+      }
+
+      #logoutButton input {
+          position: absolute;
+          top: 2em;
+          right: 2em;
+      }
+    </style>
+</head>
+<body>
+    <h2>Error 403: AuthorizationFailed.</h2>
+    <h:panelGroup id="logoutButton" layout="block">
+        <FORM METHOD="POST" ACTION="ibm_security_logout" NAME="logout">
+          <input type="submit" name="logout" value="Log Out" />
+          <input type="HIDDEN" name="logoutExitPage" value="/home" />
+        </FORM>
+    </h:panelGroup>
+    <hr/>
+    <p>You are not authorized to use this application.</p>
+</body>
+</html>
